requirements.txt
flask==0.12.2
requests
black
flask-ngrok
pillow
torch